medical-kg-drug-repurposing/requirements.txt
# Medical Knowledge Graph - Drug Repurposing
# Python 3.11+ required

# ============================================
# Core Scientific Computing
# ============================================
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ============================================
# Machine Learning - PyTorch
# ============================================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ============================================
# Graph Neural Networks
# ============================================
torch-geometric>=2.4.0
# torch-scatter and torch-sparse will be installed separately after torch

# ============================================
# NLP Libraries
# ============================================
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
spacy>=3.5.0
scispacy>=0.5.3

# SciSpacy models (will be downloaded separately)
# en_core_sci_sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_sm-0.5.3.tar.gz

# ============================================
# Graph Database
# ============================================
neo4j>=5.9.0
py2neo>=2021.2.3

# ============================================
# Network Analysis
# ============================================
networkx>=3.1
pyvis>=0.3.2

# ============================================
# Web Framework
# ============================================
streamlit>=1.24.0

# ============================================
# Data Collection & Processing
# ============================================
requests>=2.31.0
beautifulsoup4>=4.12.0
biopython>=1.81
lxml>=4.9.0

# ============================================
# Development Tools
# ============================================
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
ipykernel>=6.23.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# ============================================
# Optional: For advanced visualizations
# ============================================
graphviz>=0.20.0
pydot>=1.4.0
